+++ tests/unit/resources/test_factory.py
@@ -11,7 +11,7 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from botocore.model import ServiceModel, StructureShape
+from botocore.model import DenormalizedStructureBuilder, ServiceModel
 from boto3.exceptions import ResourceLoadException
 from boto3.resources.base import ServiceResource
 from boto3.resources.collection import CollectionManager
@@ -126,11 +126,14 @@ class TestResourceFactory(BaseTestResourceFactory):
                 }
             }
         }
-        shape = mock.Mock()
-        shape.members = {
-            'ETag': None,
-            'LastModified': None,
-        }
+        shape = DenormalizedStructureBuilder().with_members({
+            'ETag': {
+                'type': 'string',
+            },
+            'LastModified': {
+                'type': 'string'
+            }
+        }).build_model()
         service_model = mock.Mock()
         service_model.shape_for.return_value = shape
 
@@ -357,12 +360,20 @@ class TestResourceFactory(BaseTestResourceFactory):
                 }
             }
         }
-        shape = mock.Mock()
-        shape.members = {
-            'Url': None,
-            'ETag': None,
-            'LastModified': None,
-        }
+        shape = DenormalizedStructureBuilder().with_members({
+            'ETag': {
+                'type': 'string',
+                'shape_name': 'ETag'
+            },
+            'LastModified': {
+                'type': 'string',
+                'shape_name': 'LastModified'
+            },
+            'Url': {
+                'type': 'string',
+                'shape_name': 'Url'
+            }
+        }).build_model()
         service_model = mock.Mock()
         service_model.shape_for.return_value = shape
 
@@ -401,12 +412,17 @@ class TestResourceFactory(BaseTestResourceFactory):
             # Note the lack of a `load` method. These resources
             # are usually loaded via a call on a parent resource.
         }
-        shape = mock.Mock()
-        shape.members = {
-            'Url': None,
-            'ETag': None,
-            'LastModified': None,
-        }
+        shape = DenormalizedStructureBuilder().with_members({
+            'ETag': {
+                'type': 'string',
+            },
+            'LastModified': {
+                'type': 'string'
+            },
+            'Url': {
+                'type': 'string'
+            }
+        }).build_model()
         service_model = mock.Mock()
         service_model.shape_for.return_value = shape
 
@@ -517,7 +533,7 @@ class TestResourceFactory(BaseTestResourceFactory):
             'Queue': {}
         }
         service_model = ServiceModel({})
-        mock_model.return_value.name = 'Queues'
+        mock_model.return_value.name = 'queues'
 
         resource = self.load('test', 'test', model, defs, service_model)()